NVC-12
- Criada a entidade Terceiro;
- Criada a classe de serviço de Terceiro;
- Criada a Controller de Terceiro;
- Criado relacionamento unidirecional em PedidoVenda com Terceiro;
- Alterado relacionamento bidirecional entre PedidoVenda e Produtos para
unidirecional;
- Alteração no enum Movimentacao;
- Criado um mapper para PedidoVenda;
- Criado os DTOs de PedidoVenda e Terceiro.

+++ src/main/java/br/com/dev/ecommerce/estoque/enums/Movimentacao.java
@@ -2,25 +2,22 @@ package br.com.dev.ecommerce.estoque.enums;
 
 public enum Movimentacao {
 
-	SALDO_INICIAL(0, "Saldo Inicial"),
-	ENTRADA(1, "Entrada"),
-	SAIDA(2, "Saida"),
-	TRANSFERENCIA(3, "Transferencia");
-	
-	private int id;
+	SALDO_INICIAL(0L, "Saldo Inicial"), ENTRADA(1L, "Entrada"), SAIDA(2L, "Saida"), TRANSFERENCIA(3L, "Transferencia");
+
+	private Long id;
 
 	private String tipo;
 
-	private Movimentacao(int id, String tipo) {
+	private Movimentacao(Long id, String tipo) {
 		this.id = id;
 		this.tipo = tipo;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
